Keep uwsgi plugin from crashing the agent
Uwsgi plugin sometimes receives truncated data back from the uwsgi
stats server. When this happens, json.loads() fails to process the
data, and throws an exception. This bubbled up all the way to the
main python process and halted the agent.

IMO, this should not be an agent-killing problem. Wrapping json.loads()
with a try-except block and giving it some default behavior keeps it
from exploding when it recieves bad data.